🚀 Massive Refactor & Major Upgrade: Introducing 🎶 Svara – reborn as a cleaner, faster & more polished music player
- Broke down the previously bloated `main.py` into **two structured components**:
  - 🎨 **UI Component** – clean separation of the interface, visuals, and user interactions
  - 🎵 **Logic Component** – all playback, control handling, and VLC-related functionality
- 🗑 Removed unused/legacy scripts and redundant files → repository is lighter and more maintainable
- 🔄 Unified project entry point by merging the old splash screen entry into `main.py` → reduced complexity and easier debugging

---

- 🎉 Officially rebranded the app as **Svara** (Sanskrit for "musical notes") → better identity and cultural touch
- 🖼 Introduced a **new professional logo** for the app
- 🎬 Replaced the outdated splash screen GIF with a **new animated splash** based on the new logo → more modern, premium, and engaging app startup experience

---

- 📜 **Auto-scrolling tracklist**: when a song starts playing, the tracklist scrolls automatically to bring the current track to the top.
  → This ensures users always see the song in context without having to scroll manually.

---

- 🧹 **CustomTkinter Conflicts**: removed multiple exception cases caused by conflicting widget states
- 🎯 **Tracklist Hover Issues**:
  - Fixed bug where hovering highlighted *every* song at once
  - Fixed hover only triggering at the **edges of track rows** → now highlights properly when hovering anywhere on the song row
- 🖱 **Scrolling Problems**:
  - Fixed auto-scroll trigger when hovering over tracklist → now behaves as expected
  - Restored mouse wheel scrolling (previously broken)
  - Increased scroll speed to make navigation through long playlists more natural and responsive
- 🔊 **VLC Volume Sync**:
  - Fixed major issue where playback always started at **100% volume** while the UI showed 50%
  - Volume slider now initializes correctly on app start → actual playback volume matches UI display

---

- ✅ Refactored architecture = cleaner separation of concerns, better readability, and easier future feature expansion
- 🧹 Removed unused files and redundant logic → reduced repo size and improved performance
- 🚀 Overall, the app feels lighter, faster, and smoother thanks to better code structure and fewer runtime conflicts

---

This release was made possible thanks to the **combined effort and collaboration** of:
- **Swapnanil Basak** 🎶
- **Piyush** 🎶

Both worked jointly on:
- Refactoring and splitting the codebase into UI & Logic components
- Designing and integrating the new logo + splash screen animation
- Fixing major UI/UX issues (hover glitches, scrolling bugs, volume sync problems)
- Removing unused files, merging entry points, and cleaning code for better maintainability
- Testing, debugging, and optimizing performance for smoother user experience

---

💡 After weeks of careful refactoring, debugging late into the night, and polishing the user experience, **Svara has been reborn**.
This is not just an update — it is a **milestone release** that makes the app cleaner, faster, and way more enjoyable to use. 🚀
